Day 5: Binary boarding
Part 1: Calculates the seatId for each of the seats and identifies the maximum seat Id in the input. The seatId = row * 8 + col.
Part 2: Since the flight is full, and only first few seatIds and last few seatIds are empty, my seat has to be between two other seatIds. Stores all seatIds in a map, iterate from 0 to 1024, and find the missing seatId such that seatId + 1 and seatId - 1 are present.