Day6: Custom customs
Part1: For each group, identify the total number of questions to which anyone in the group answered 'yes', and return the sum of these counts.
Part2: For each group, identify the total number of questions to which everyone in the group answered 'yes', and return the sum of these counts.