Day8: Handheld halting
Part1: Find the value of the accumulator when any instruction gets executed a second time.
Part2: Find the value of the accumulator when we change either a 'jmp' instruction to a 'nop', or a 'nop' instruction to a 'jmp' such that, it results in the instruction execution reaching the end.