Day9: Encoding Error
Part1: Get the first invalid number in the given list of numbers. A number is invalid if it is not the sum of two unique numbers from the 25 numbers preceeding it.
Part2: Find a contiguous set of at least two numbers which add up to the invalid number, and then return the sum of the min and max numbers from this contiguous set.